Update scripts.js
Instant Finish or Sudden Death button logic added. Player object was also added to hold properties and methods that pertain to the current game player. The idea for the player object was implemented as an attempt to implement scoring up or down for the game player, depending on whether the right guess or wrong guess was provided.